refactor: Extract the in-process tRPC handler link from the test helpers
The terminating mock link that resolves procedures from a handler map moves
from tests/utils into src/lib/trpc/handler-link.ts so the public demo can run
the real tRPC client against canned data with the same link the component
tests use. TRPCProvider takes an optional links override for that.
