+++ src/lib/trpc/provider.tsx
@@ -9,8 +9,9 @@
 
 import { useState, type ReactNode } from "react";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { httpBatchLink } from "@trpc/client";
+import { httpBatchLink, type TRPCLink } from "@trpc/client";
 import superjson from "superjson";
+import type { AppRouter } from "@/server/trpc/root";
 import { trpc } from "./client";
 import { getQueryClient } from "./query-client";
 
@@ -33,7 +34,29 @@ function getBaseUrl() {
   return `http://localhost:${process.env.PORT ?? 3000}`;
 }
 
+function createHttpLinks(): TRPCLink<AppRouter>[] {
+  return [
+    httpBatchLink({
+      url: `${getBaseUrl()}/api/trpc`,
+      transformer: superjson,
+      // Include credentials for cookie-based auth
+      fetch(url, options) {
+        return fetch(url, {
+          ...options,
+          credentials: "include",
+        });
+      },
+    }),
+  ];
+}
+
 interface TRPCProviderProps {
+  /**
+   * Link chain override. Defaults to the batching HTTP link against
+   * `/api/trpc`; the public demo passes an in-process handler link
+   * (`createHandlerLink`) so the same hooks resolve from canned data.
+   */
+  links?: TRPCLink<AppRouter>[];
   children: ReactNode;
 }
 
@@ -49,29 +72,13 @@ interface TRPCProviderProps {
  * </TRPCProvider>
  * ```
  */
-export function TRPCProvider({ children }: TRPCProviderProps) {
+export function TRPCProvider({ links, children }: TRPCProviderProps) {
   // Use the shared QueryClient from query-client.ts
   // This ensures server prefetching and client components use the same instance
   // during SSR, preventing hydration mismatches.
   const queryClient = getQueryClient();
 
-  const [trpcClient] = useState(() =>
-    trpc.createClient({
-      links: [
-        httpBatchLink({
-          url: `${getBaseUrl()}/api/trpc`,
-          transformer: superjson,
-          // Include credentials for cookie-based auth
-          fetch(url, options) {
-            return fetch(url, {
-              ...options,
-              credentials: "include",
-            });
-          },
-        }),
-      ],
-    })
-  );
+  const [trpcClient] = useState(() => trpc.createClient({ links: links ?? createHttpLinks() }));
 
   return (
     <trpc.Provider client={trpcClient} queryClient={queryClient}>
